requirements.txt
requests
tkinter 
python-vlc